moved q2-demux changes into multiplexed sequences
