+++ q2_types/per_sample_sequences/__init__.py
@@ -23,22 +23,11 @@ from ._format import (CasavaOneEightSingleLanePerSampleDirFmt,
                       PairedEndFastqManifestPhred33V2,
                       PairedEndFastqManifestPhred64V2,
                       QIIME1DemuxFormat, QIIME1DemuxDirFmt,
-                      SampleIdIndexedSingleEndPerSampleDirFmt,
-                      EMPMultiplexedDirFmt,
-                      ErrorCorrectionDetailsFmt,
-                      ErrorCorrectionDetailsDirFmt,
-                      EMPSingleEndDirFmt,
-                      EMPSingleEndCasavaDirFmt,
-                      EMPPairedEndDirFmt,
-                      EMPPairedEndCasavaDirFmt)
+                      SampleIdIndexedSingleEndPerSampleDirFmt)
 
 from ._type import (Sequences, SequencesWithQuality,
                     PairedEndSequencesWithQuality,
-                    JoinedSequencesWithQuality,
-                    RawSequences,
-                    EMPSingleEndSequences,
-                    EMPPairedEndSequences,
-                    ErrorCorrectionDetails)
+                    JoinedSequencesWithQuality)
 
 __all__ = ['CasavaOneEightSingleLanePerSampleDirFmt',
            'CasavaOneEightLanelessPerSampleDirFmt',
@@ -53,12 +42,6 @@ __all__ = ['CasavaOneEightSingleLanePerSampleDirFmt',
            'SingleEndFastqManifestPhred64V2',
            'PairedEndFastqManifestPhred33V2',
            'PairedEndFastqManifestPhred64V2', 'QIIME1DemuxFormat',
-           'QIIME1DemuxDirFmt', 'SampleIdIndexedSingleEndPerSampleDirFmt',
-           'RawSequences', 'EMPSingleEndSequences', 'EMPPairedEndSequences',
-           'EMPMultiplexedDirFmt',
-           'ErrorCorrectionDetails', 'ErrorCorrectionDetailsFmt',
-           'ErrorCorrectionDetailsDirFmt', 'EMPSingleEndDirFmt',
-           'EMPSingleEndCasavaDirFmt', 'EMPPairedEndDirFmt',
-           'EMPPairedEndCasavaDirFmt']
+           'QIIME1DemuxDirFmt', 'SampleIdIndexedSingleEndPerSampleDirFmt']
 
 importlib.import_module('q2_types.per_sample_sequences._transformer')
