+++ q2_types/per_sample_sequences/_type.py
@@ -11,8 +11,7 @@ from qiime2.plugin import SemanticType
 from ..plugin_setup import plugin
 from ..sample_data import SampleData
 from . import (QIIME1DemuxDirFmt, SingleLanePerSampleSingleEndFastqDirFmt,
-               SingleLanePerSamplePairedEndFastqDirFmt, EMPSingleEndDirFmt,
-               EMPPairedEndDirFmt, ErrorCorrectionDetailsDirFmt)
+               SingleLanePerSamplePairedEndFastqDirFmt)
 
 
 Sequences = SemanticType('Sequences', variant_of=SampleData.field['type'])
@@ -22,17 +21,10 @@ PairedEndSequencesWithQuality = SemanticType(
     'PairedEndSequencesWithQuality', variant_of=SampleData.field['type'])
 JoinedSequencesWithQuality = SemanticType(
     'JoinedSequencesWithQuality', variant_of=SampleData.field['type'])
-RawSequences = SemanticType('RawSequences')
-EMPSingleEndSequences = SemanticType('EMPSingleEndSequences')
-EMPPairedEndSequences = SemanticType('EMPPairedEndSequences')
-ErrorCorrectionDetails = SemanticType('ErrorCorrectionDetails')
 
 plugin.register_semantic_types(Sequences, SequencesWithQuality,
                                PairedEndSequencesWithQuality,
-                               JoinedSequencesWithQuality,
-                               RawSequences, EMPSingleEndSequences,
-                               EMPPairedEndSequences,
-                               ErrorCorrectionDetails)
+                               JoinedSequencesWithQuality)
 
 plugin.register_artifact_class(
     SampleData[Sequences],
@@ -59,24 +51,4 @@ plugin.register_artifact_class(
     description=("Collections of unjoined paired-end sequences with quality "
                  "scores associated with specified samples (i.e., "
                  "demultiplexed sequences).")
-)
-# TODO: remove when aliasing exists
-plugin.register_semantic_type_to_format(
-    RawSequences,
-    artifact_format=EMPSingleEndDirFmt
-)
-
-plugin.register_semantic_type_to_format(
-    EMPSingleEndSequences,
-    artifact_format=EMPSingleEndDirFmt
-)
-
-plugin.register_semantic_type_to_format(
-    EMPPairedEndSequences,
-    artifact_format=EMPPairedEndDirFmt
-)
-
-plugin.register_semantic_type_to_format(
-    ErrorCorrectionDetails,
-    artifact_format=ErrorCorrectionDetailsDirFmt
 )
